+++ packages/rax-pwa-webpack-plugin/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "css-loader": "^2.1.1",
+    "html-minifier": "^4.0.0",
     "loader-utils": "^1.2.3",
     "path-to-regexp": "^3.0.0",
     "postcss-loader": "^3.0.0",
