+++ packages/rax-pwa-webpack-plugin/src/config/RaxPWAPlugins.js
@@ -1,0 +1,4 @@
+module.exports = [
+  require('../plugins/RaxDocumentPlugin'),
+  require('../plugins/RaxSPAPlugin')
+];
