+++ packages/rax-pwa-webpack-plugin/src/index.js
@@ -1,7 +1,9 @@
 const ClientLoader = require('./ClientLoader');
 const ServerlessLoader = require('./ServerlessLoader');
+const RaxPWAPlugin = require('./RaxPWAPlugin');
 
 module.exports = {
   ClientLoader,
-  ServerlessLoader
+  ServerlessLoader,
+  RaxPWAPlugin
 };
