+++ packages/rax-pwa-webpack-plugin/src/utils/isProductionMode.js
@@ -1,0 +1,5 @@
+const isProductionMode = (compiler) => {
+  return compiler.options.mode === 'production' || !compiler.options.mode;
+};
+
+module.exports = isProductionMode;
